test: 프로덕션 코드와 테스트 클래스 간의 패키지 경로 통일
