+++ src/test/java/domain/movement/MoveAmountTest.java
@@ -1,9 +1,8 @@
-package domain.direction;
+package domain.movement;
 
 import static domain.util.AssertUtils.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.movement.MoveAmount;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
